requirements.txt
fastapi
uvicorn
google-generativeai
ollama
rich
tenacity
